Print Memory Copies due to Omp Target Clauses
An Analysis pass, that prints the host/device memory copies introduced due to the target offloading clauses used by the programmer.
It follows the Omp Spec 5.0
This is a global/module level pass, that does interprocedural analysis to interpret the semantics of all the target offloading clauses.
First we analyze every function, and parse the arguments being passed to the Omp RTL calls, and extract the variables mapped, the map type and the array section.
Then we perform a global analysis to do a BFS traversal of every function basic block, and maintain the reference count, and interpret when the variable is copied from host to device and viceversa.

Checked in two lit test cases from the DRACC benchmark.
Use the following pass to test this analysis pass.
opt -S -passes='print-omp-diagnostics'